Fix bug with loading mean

+++ datasets.py
@@ -46,7 +46,7 @@ def _get_mnist(batch_size, split="train", binarized=None, repeat=True, shuffle=T
 
   path = os.path.join(data_dir, split + ".npy")
   np_ims = np.load(path)
-  mean = np.load(os.path.join(data_dir, "train_mean.npy"))
+  mean = np.load(os.path.join(data_dir, "train_mean.npy")).astype(np.float32)
   dataset = tf.data.Dataset.from_tensor_slices(np_ims)
 
   if binarized == "dynamic":
